requirements.txt
streamlit>=1.28.0
spacy>=3.7.0
neo4j>=5.14.0
PyPDF2>=3.0.1
pdfplumber>=0.10.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
langchain-openai>=0.0.5
faiss-cpu>=1.7.4
huggingface-hub>=0.19.0
neo4j-graphrag>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
